refactor(import-core): drop import-engine buildPersistableEntity shim
Remove the cross-module re-export in `import-engine.ts` and route the
package barrel directly to `import-engine.helpers.js`. Migrate the one
core test consumer to import from the helpers module. Update stale
JSDoc that pointed through the now-removed shim.

The import-sp engine maintains its own `buildPersistableEntity`
re-export from `orchestrator-helpers.js` — that's a separate module
boundary and out of scope for this PR.

Pre-production policy: no aliases, shims, or re-exports.
